Revert CSS filter utils
