--- src/generators/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './inversion';
-export * from './matrix';
